requirements.txt
# Core LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# Mistral Integration
langchain-mistralai>=0.1.0
openai>=1.30.0

# Tavily Search Tool
tavily-python>=0.3.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Data handling
pandas>=2.0.0

# Token utils
tiktoken>=0.6.0

# Logging
rich>=13.7.0

# Retry
tenacity>=8.2.0

# JSON handling
orjson>=3.9.0

# Typing
pydantic>=2.5.0

# Optional parsing
html5lib>=1.1